Move completion tests
