+++ src/Analysis/Engine/Test/LanguageServerTests.cs
@@ -32,7 +32,6 @@ using Microsoft.PythonTools.Analysis.FluentAssertions;
 using Microsoft.PythonTools.Analysis.Infrastructure;
 using Microsoft.PythonTools.Interpreter;
 using Microsoft.PythonTools.Interpreter.Ast;
-using Microsoft.PythonTools.Parsing.Ast;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using TestUtilities;
 
@@ -224,454 +223,6 @@ namespace AnalysisTests {
         }
 
         [TestMethod, Priority(0)]
-        public async Task TopLevelCompletions() {
-            var s = await CreateServer(TestData.GetPath(@"TestData\AstAnalysis"));
-
-            await AssertCompletion(
-                s,
-                GetDocument(@"TestData\AstAnalysis\TopLevelCompletions.py"),
-                new[] { "x", "y", "z", "int", "float", "class", "def", "while", "in" },
-                new[] { "return", "sys", "yield" }
-            );
-
-            // Completions in function body
-            await AssertCompletion(
-                s,
-                GetDocument(@"TestData\AstAnalysis\TopLevelCompletions.py"),
-                new[] { "x", "y", "z", "int", "float", "class", "def", "while", "in", "return", "yield" },
-                new[] { "sys" },
-                position: new Position { line = 5, character = 5 }
-            );
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionInForStatement() {
-            var s = await CreateServer();
-            Uri u;
-
-            u = await AddModule(s, "for  ");
-            await AssertCompletion(s, u, new[] { "for" }, new string[0], new SourceLocation(1, 4));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 5));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "for  x ");
-            await AssertCompletion(s, u, new[] { "for" }, new string[0], new SourceLocation(1, 4));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 5));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 6));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 7));
-            await AssertCompletion(s, u, new[] { "in" }, new[] { "for", "abs" }, new SourceLocation(1, 8));
-            await s.UnloadFileAsync(u);
-
-            // TODO: Fix parser to parse "for x i" as ForStatement and not ForStatement+ExpressionStatement
-            //u = await AddModule(s, "for x i");
-            //await AssertCompletion(s, u, new[] { "in" }, new[] { "for", "abs" }, new SourceLocation(1, 8), applicableSpan: new SourceSpan(1, 7, 1, 8));
-            //await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "for x in ");
-            await AssertCompletion(s, u, new[] { "in" }, new[] { "for", "abs" }, new SourceLocation(1, 7));
-            await AssertCompletion(s, u, new[] { "in" }, new[] { "for", "abs" }, new SourceLocation(1, 9));
-            await AssertCompletion(s, u, new[] { "abs", "x" }, new string[0], new SourceLocation(1, 10));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "def f():\n    for ");
-            await AssertNoCompletion(s, u, new SourceLocation(2, 9));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "def f():\n    for x in ");
-            await AssertCompletion(s, u, new[] { "in" }, new[] { "for", "abs" }, new SourceLocation(2, 11));
-            await AssertCompletion(s, u, new[] { "in" }, new[] { "for", "abs" }, new SourceLocation(2, 13));
-            await AssertCompletion(s, u, new[] { "abs", "x" }, new string[0], new SourceLocation(2, 14));
-            await s.UnloadFileAsync(u);
-
-            if (!(this is LanguageServerTests_V2)) {
-                u = await AddModule(s, "async def f():\n    async for x in ");
-                await AssertCompletion(s, u, new[] { "async", "for" }, new string[0], new SourceLocation(2, 5));
-                await AssertCompletion(s, u, new[] { "async", "for" }, new string[0], new SourceLocation(2, 10));
-                await AssertCompletion(s, u, new[] { "async", "for" }, new string[0], new SourceLocation(2, 14));
-                await AssertNoCompletion(s, u, new SourceLocation(2, 15));
-                await s.UnloadFileAsync(u);
-            }
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionInFunctionDefinition() {
-            var s = await CreateServer();
-            var u = await AddModule(s, "def f(a, b:int, c=2, d:float=None): pass");
-
-            await AssertNoCompletion(s, u, new SourceLocation(1, 5));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 7));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 8));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 10));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 14));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 17));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 19));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 29));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 34));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 35));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 36));
-
-            if (this is LanguageServerTests_V2) {
-                u = await AddModule(s, "@dec\ndef  f(): pass");
-                await AssertAnyCompletion(s, u, new SourceLocation(1, 1));
-                await AssertCompletion(s, u, new[] { "abs" }, new[] { "def" }, new SourceLocation(1, 2));
-                await AssertCompletion(s, u, new[] { "def" }, new string[0], new SourceLocation(2, 1));
-                await AssertCompletion(s, u, new[] { "def" }, new string[0], new SourceLocation(2, 4));
-                await AssertNoCompletion(s, u, new SourceLocation(2, 5));
-                await AssertNoCompletion(s, u, new SourceLocation(2, 6));
-            } else {
-                u = await AddModule(s, "@dec\nasync   def  f(): pass");
-                await AssertAnyCompletion(s, u, new SourceLocation(1, 1));
-                await AssertCompletion(s, u, new[] { "abs" }, new[] { "def" }, new SourceLocation(1, 2));
-                await AssertCompletion(s, u, new[] { "def" }, new string[0], new SourceLocation(2, 1));
-                await AssertCompletion(s, u, new[] { "def" }, new string[0], new SourceLocation(2, 12));
-                await AssertNoCompletion(s, u, new SourceLocation(2, 13));
-                await AssertNoCompletion(s, u, new SourceLocation(2, 14));
-            }
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionInClassDefinition() {
-            var s = await CreateServer();
-            var u = await AddModule(s, "class C(object, parameter=MC): pass");
-
-            await AssertNoCompletion(s, u, new SourceLocation(1, 8));
-            if (this is LanguageServerTests_V2) {
-                await AssertCompletion(s, u, new[] { "object" }, new[] { "metaclass=" }, new SourceLocation(1, 9));
-            } else {
-                await AssertCompletion(s, u, new[] { "metaclass=", "object" }, new string[0], new SourceLocation(1, 9));
-            }
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 15));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 17));
-            await AssertCompletion(s, u, new[] { "object" }, new[] { "metaclass=" }, new SourceLocation(1, 29));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 30));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 31));
-
-            u = await AddModule(s, "class D(o");
-            await AssertNoCompletion(s, u, new SourceLocation(1, 8));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 9));
-
-            u = await AddModule(s, "class E(metaclass=MC,o): pass");
-            await AssertCompletion(s, u, new[] { "object" }, new[] { "metaclass=" }, new SourceLocation(1, 22));
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionInWithStatement() {
-            var s = await CreateServer();
-            Uri u;
-
-            u = await AddModule(s, "with x as y, z as w: pass");
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 6));
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "abs", "dir" }, new SourceLocation(1, 8));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 11));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 14));
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "abs", "dir" }, new SourceLocation(1, 17));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 20));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 21));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "with ");
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 6));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "with x ");
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 6));
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "abs", "dir" }, new SourceLocation(1, 8));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "with x as ");
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 6));
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "abs", "dir" }, new SourceLocation(1, 8));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 11));
-            await s.UnloadFileAsync(u);
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionInImport() {
-            var s = await CreateServer();
-            var u = await AddModule(s, "import unittest.case as C, unittest\nfrom unittest.case import TestCase as TC, TestCase");
-
-            await AssertCompletion(s, u, new[] { "from", "import", "abs", "dir" }, new[] { "abc" }, new SourceLocation(1, 7));
-            await AssertCompletion(s, u, new[] { "abc", "unittest" }, new[] { "abs", "dir" }, new SourceLocation(1, 8));
-            await AssertCompletion(s, u, new[] { "case" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(1, 17));
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(1, 22));
-            await AssertNoCompletion(s, u, new SourceLocation(1, 25));
-            await AssertCompletion(s, u, new[] { "abc", "unittest" }, new[] { "abs", "dir" }, new SourceLocation(1, 28));
-
-            await AssertCompletion(s, u, new[] { "from", "import", "abs", "dir" }, new[] { "abc" }, new SourceLocation(2, 5));
-            await AssertCompletion(s, u, new[] { "abc", "unittest" }, new[] { "abs", "dir" }, new SourceLocation(2, 6));
-            await AssertCompletion(s, u, new[] { "case" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(2, 15));
-            await AssertCompletion(s, u, new[] { "import" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(2, 20));
-            await AssertCompletion(s, u, new[] { "import" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(2, 22), applicableSpan: new SourceSpan(2, 20, 2, 26));
-            await AssertCompletion(s, u, new[] { "TestCase" }, new[] { "abs", "dir", "case" }, new SourceLocation(2, 27));
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(2, 36));
-            await AssertNoCompletion(s, u, new SourceLocation(2, 39));
-            await AssertCompletion(s, u, new[] { "TestCase" }, new[] { "abs", "dir", "case" }, new SourceLocation(2, 44));
-
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "from unittest.case imp\n\npass");
-            await AssertCompletion(s, u, new[] { "import" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(1, 22), applicableSpan: new SourceSpan(1, 20, 1, 23));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "import unittest.case a\n\npass");
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(1, 23), applicableSpan: new SourceSpan(1, 22, 1, 23));
-            await s.UnloadFileAsync(u);
-            u = await AddModule(s, "from unittest.case import TestCase a\n\npass");
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "abc", "unittest", "abs", "dir" }, new SourceLocation(1, 37), applicableSpan: new SourceSpan(1, 36, 1, 37));
-            await s.UnloadFileAsync(u);
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionForOverride() {
-            var s = await CreateServer();
-            var u = await AddModule(s, "class A(object):\n    def i(): pass\n    def \npass");
-
-            await AssertNoCompletion(s, u, new SourceLocation(2, 9));
-            await AssertCompletion(s, u, new[] { "def" }, new[] { "__init__" }, new SourceLocation(3, 8));
-            await AssertCompletion(s, u, new[] { "__init__" }, new[] { "def" }, new SourceLocation(3, 9), cmpKey: ci => ci.label);
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionForOverrideArgs() {
-            using (var s = await CreateServer()) {
-                var u = await AddModule(s, "class A:\n    def bar(arg=None): pass\n\nclass B(A):\n    def b");
-
-                await AssertNoCompletion(s, u, new SourceLocation(2, 9));
-                await AssertCompletion(s, u, 
-                    new[] { "bar(arg=None):\r\n    return super().bar()" },
-                    new[] { "bar(arg = None):\r\n    return super().bar()" }, 
-                    new SourceLocation(5, 10));
-            }
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionInDecorator() {
-            var s = await CreateServer();
-            var u = await AddModule(s, "@dec\ndef f(): pass\n\nx = a @ b");
-
-            await AssertCompletion(s, u, new[] { "f", "x", "property", "abs" }, new[] { "def" }, new SourceLocation(1, 2));
-            await AssertCompletion(s, u, new[] { "f", "x", "property", "abs" }, new[] { "def" }, new SourceLocation(4, 8));
-
-            u = await AddModule(s, "@");
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 2));
-
-            u = await AddModule(s, "import unittest\n\n@unittest.\n");
-            await AssertCompletion(s, u, new[] { "TestCase", "skip", "skipUnless" }, new[] { "abs", "def" }, new SourceLocation(3, 11));
-
-            u = await AddModule(s, "import unittest\n\n@unittest.\ndef f(): pass");
-            await AssertCompletion(s, u, new[] { "TestCase", "skip", "skipUnless" }, new[] { "abs", "def" }, new SourceLocation(3, 11));
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionInRaise() {
-            var s = await CreateServer();
-            var u = await AddModule(s, "raise ");
-            await AssertCompletion(s, u, new[] { "Exception", "ValueError" }, new[] { "def", "abs" }, new SourceLocation(1, 7));
-            await s.UnloadFileAsync(u);
-
-            if (!(this is LanguageServerTests_V2)) {
-                u = await AddModule(s, "raise Exception from ");
-                await AssertCompletion(s, u, new[] { "Exception", "ValueError" }, new[] { "def", "abs" }, new SourceLocation(1, 7));
-                await AssertCompletion(s, u, new[] { "from" }, new[] { "Exception", "def", "abs" }, new SourceLocation(1, 17));
-                await AssertAnyCompletion(s, u, new SourceLocation(1, 22));
-                await s.UnloadFileAsync(u);
-
-                u = await AddModule(s, "raise Exception fr");
-                await AssertCompletion(s, u, new[] { "from" }, new[] { "Exception", "def", "abs" }, new SourceLocation(1, 19), applicableSpan: new SourceSpan(1, 17, 1, 19));
-                await s.UnloadFileAsync(u);
-            }
-
-            u = await AddModule(s, "raise Exception, x, y");
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 17));
-            await AssertAnyCompletion(s, u, new SourceLocation(1, 20));
-            await s.UnloadFileAsync(u);
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionInExcept() {
-            var s = await CreateServer();
-            Uri u;
-            u = await AddModule(s, "try:\n    pass\nexcept ");
-            await AssertCompletion(s, u, new[] { "Exception", "ValueError" }, new[] { "def", "abs" }, new SourceLocation(3, 8));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "try:\n    pass\nexcept (");
-            await AssertCompletion(s, u, new[] { "Exception", "ValueError" }, new[] { "def", "abs" }, new SourceLocation(3, 9));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "try:\n    pass\nexcept Exception  as ");
-            await AssertCompletion(s, u, new[] { "Exception", "ValueError" }, new[] { "def", "abs" }, new SourceLocation(3, 8));
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "Exception", "def", "abs" }, new SourceLocation(3, 18));
-            await AssertNoCompletion(s, u, new SourceLocation(3, 22));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "try:\n    pass\nexc");
-            await AssertCompletion(s, u, new[] { "except", "def", "abs" }, new string[0], new SourceLocation(3, 3));
-            await s.UnloadFileAsync(u);
-
-            u = await AddModule(s, "try:\n    pass\nexcept Exception a");
-            await AssertCompletion(s, u, new[] { "as" }, new[] { "Exception", "def", "abs" }, new SourceLocation(3, 19), applicableSpan: new SourceSpan(3, 18, 3, 19));
-            await s.UnloadFileAsync(u);
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionAfterDot() {
-            var s = await CreateServer();
-            Uri u;
-
-            u = await AddModule(s, "x = 1\nx. n\nx.(  )\nx(x.  )\nx.  \nx  ");
-            await AssertCompletion(s, u, new[] { "real", "imag" }, new[] { "abs" }, new SourceLocation(2, 3));
-            await AssertCompletion(s, u, new[] { "real", "imag" }, new[] { "abs" }, new SourceLocation(2, 4));
-            await AssertCompletion(s, u, new[] { "real", "imag" }, new[] { "abs" }, new SourceLocation(2, 5));
-            await AssertCompletion(s, u, new[] { "real", "imag" }, new[] { "abs" }, new SourceLocation(3, 3));
-            await AssertCompletion(s, u, new[] { "real", "imag" }, new[] { "abs" }, new SourceLocation(4, 5));
-            await AssertCompletion(s, u, new[] { "real", "imag" }, new[] { "abs" }, new SourceLocation(5, 4));
-            await AssertCompletion(s, u, new[] { "abs" }, new[] { "real", "imag" }, new SourceLocation(6, 2));
-            await AssertNoCompletion(s, u, new SourceLocation(6, 3));
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionAfterAssign() {
-            var s = await CreateServer();
-            Uri u;
-
-            u = await AddModule(s, "x = x\ny = ");
-            await AssertCompletion(s, u, new[] { "x", "abs" }, null, new SourceLocation(1, 5));
-            await AssertCompletion(s, u, new[] { "x", "abs" }, null, new SourceLocation(2, 5));
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionWithNewDot() {
-            // LSP assumes that the text buffer is up to date with typing,
-            // which means the language server must know about dot for a
-            // dot completion.
-            // To do this, we have to support using a newer tree than the
-            // current analysis, so that we can quickly parse the new text
-            // with the dot but not block on reanalysis.
-            var s = await CreateServer();
-            var code = @"
-class MyClass:
-    def f(self): pass
-
-mc = MyClass()
-mc
-";
-            int testLine = 5;
-            int testChar = 2;
-
-            var mod = await AddModule(s, code);
-
-            // Completion after "mc " should normally be blank
-            await AssertCompletion(s, mod,
-                new string[0],
-                new string[0],
-                position: new Position { line = testLine, character = testChar + 1 }
-            );
-
-            // While we're here, test with the special override field
-            await AssertCompletion(s, mod,
-                new[] { "f" },
-                new[] { "abs", "bin", "int", "mc" },
-                position: new Position { line = testLine, character = testChar + 1 },
-                expr: "mc"
-            );
-
-            // Send the document update.
-            await s.DidChangeTextDocument(new DidChangeTextDocumentParams {
-                textDocument = new VersionedTextDocumentIdentifier { uri = mod, version = 1 },
-                contentChanges = new[] { new TextDocumentContentChangedEvent {
-                    text = ".",
-                    range = new Range {
-                        start = new Position { line = testLine, character = testChar },
-                        end = new Position { line = testLine, character = testChar }
-                    }
-                } },
-                // Suppress reanalysis to avoid a race
-                _enqueueForAnalysis = false
-            }, CancellationToken.None);
-
-            // Now with the "." event sent, we should see this as a dot completion
-            await AssertCompletion(s, mod,
-                new[] { "f" },
-                new[] { "abs", "bin", "int", "mc" },
-                position: new Position { line = testLine, character = testChar + 1 }
-            );
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionAfterLoad() {
-            var s = await CreateServer();
-            var mod1 = await AddModule(s, "import mod2\n\nmod2.", "mod1");
-
-            await AssertCompletion(s, mod1,
-                position: new Position { line = 2, character = 5 },
-                contains: new string[0],
-                excludes: new[] { "value" }
-            );
-
-            var mod2 = await AddModule(s, "value = 123", "mod2");
-
-            await AssertCompletion(s, mod1,
-                position: new Position { line = 2, character = 5 },
-                contains: new[] { "value" },
-                excludes: new string[0]
-            );
-
-            await s.UnloadFileAsync(mod2);
-            await s.WaitForCompleteAnalysisAsync(CancellationToken.None);
-
-            await AssertCompletion(s, mod1,
-                position: new Position { line = 2, character = 5 },
-                contains: new string[0],
-                excludes: new[] { "value" }
-            );
-        }
-
-
-        class TestCompletionHookProvider : ILanguageServerExtensionProvider {
-            public Task<ILanguageServerExtension> CreateAsync(IPythonLanguageServer server, IReadOnlyDictionary<string, object> properties, CancellationToken cancellationToken) {
-                return Task.FromResult<ILanguageServerExtension>(new TestCompletionHook());
-            }
-        }
-
-        class TestCompletionHook : ILanguageServerExtension, ICompletionExtension {
-            public void Dispose() { }
-
-            #region ILanguageServerExtension
-            public string Name => "Test completion extension";
-            public Task Initialize(IServiceContainer services, CancellationToken token) => Task.CompletedTask;
-            public Task<IReadOnlyDictionary<string, object>> ExecuteCommand(string command, IReadOnlyDictionary<string, object> properties, CancellationToken token)
-                => Task.FromResult<IReadOnlyDictionary<string, object>>(null);
-            #endregion
-
-            #region ICompletionExtension
-            public Task HandleCompletionAsync(Uri documentUri, IModuleAnalysis analysis, PythonAst tree, SourceLocation location, CompletionList completions, CancellationToken token) {
-                Assert.IsNotNull(tree);
-                Assert.IsNotNull(analysis);
-                for (int i = 0; i < completions.items.Length; ++i) {
-                    completions.items[i].insertText = "*" + completions.items[i].insertText;
-                }
-                return Task.CompletedTask;
-            }
-            #endregion
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task CompletionHook() {
-            var s = await CreateServer();
-            var u = await AddModule(s, "x = 123\nx.");
-
-            await AssertCompletion(s, u, new[] { "real", "imag" }, new string[0], new Position { line = 1, character = 2 });
-
-            await s.LoadExtensionAsync(new PythonAnalysisExtensionParams {
-                assembly = typeof(TestCompletionHookProvider).Assembly.FullName,
-                typeName = typeof(TestCompletionHookProvider).FullName
-            }, null, CancellationToken.None);
-
-            await AssertCompletion(s, u, new[] { "*real", "*imag" }, new[] { "real" }, new Position { line = 1, character = 2 });
-        }
-
-        [TestMethod, Priority(0)]
         public async Task SignatureHelp() {
             var s = await CreateServer();
             var mod = await AddModule(s, @"f()
@@ -707,33 +258,6 @@ def f(a = 2, b): pass
                 new string[] { "f(a: int)", "f(a: int=2, b: int)", "f(x: str, y: str)" },
                 new string[0]
             );
-        }
-
-        [TestMethod, Priority(0)]
-        public async Task MultiPartDocument() {
-            var s = await CreateServer();
-
-            var mod = await AddModule(s, "x = 1", "mod");
-            var modP2 = new Uri(mod, "#2");
-            var modP3 = new Uri(mod, "#3");
-
-            await AssertCompletion(s, mod, new[] { "x" }, null);
-
-            Assert.AreEqual(Tuple.Create("y = 2", 1), await ApplyChange(s, modP2, DocumentChange.Insert("y = 2", SourceLocation.MinValue)));
-            await s.WaitForCompleteAnalysisAsync(CancellationToken.None);
-
-            await AssertCompletion(s, modP2, new[] { "x", "y" }, null);
-
-            Assert.AreEqual(Tuple.Create("z = 3", 1), await ApplyChange(s, modP3, DocumentChange.Insert("z = 3", SourceLocation.MinValue)));
-            await s.WaitForCompleteAnalysisAsync(CancellationToken.None);
-
-            await AssertCompletion(s, modP3, new[] { "x", "y", "z" }, null);
-            await AssertCompletion(s, mod, new[] { "x", "y", "z" }, null);
-
-            await ApplyChange(s, mod, DocumentChange.Delete(SourceLocation.MinValue, SourceLocation.MinValue.AddColumns(5)));
-            await s.WaitForCompleteAnalysisAsync(CancellationToken.None);
-            await AssertCompletion(s, modP2, new[] { "y", "z" }, new[] { "x" });
-            await AssertCompletion(s, modP3, new[] { "y", "z" }, new[] { "x" });
         }
 
         [TestMethod, Priority(0)]
@@ -933,51 +457,9 @@ def f(a = 2, b): pass
                 .Select(d => $"{d.severity};{d.message};{d.source};{d.range.start.line};{d.range.start.character};{d.range.end.character}");
         }
 
-        public static async Task AssertCompletion(Server s, TextDocumentIdentifier document, IReadOnlyCollection<string> contains, IReadOnlyCollection<string> excludes, Position? position = null, CompletionContext? context = null, Func<CompletionItem, string> cmpKey = null, string expr = null, Range? applicableSpan = null) {
-            var res = await s.Completion(new CompletionParams {
-                textDocument = document,
-                position = position ?? new Position(),
-                context = context,
-                _expr = expr
-            }, CancellationToken.None);
-            DumpDetails(res);
-
-            cmpKey = cmpKey ?? (c => c.insertText);
-            var items = res.items?.Select(cmpKey).ToList() ?? new List<string>();
-
-            if (contains != null && contains.Any()) {
-                items.Should().Contain(contains);
-            }
-
-            if (excludes != null && excludes.Any()) {
-                items.Should().NotContain(excludes);
-            }
-
-            if (applicableSpan.HasValue) {
-                res._applicableSpan.Should().Be(applicableSpan);
-            }
-        }
-
         private static void DumpDetails(CompletionList completions) {
             var span = ((SourceSpan?)completions._applicableSpan) ?? SourceSpan.None;
             Debug.WriteLine($"Completed {completions._expr ?? "(null)"} at {span}");
-        }
-
-        private static async Task AssertAnyCompletion(Server s, TextDocumentIdentifier document, Position position) {
-            var res = await s.Completion(new CompletionParams { textDocument = document, position = position }, CancellationToken.None);
-            DumpDetails(res);
-            if (res.items == null || !res.items.Any()) {
-                Assert.Fail("Completions were not returned");
-            }
-        }
-
-        private static async Task AssertNoCompletion(Server s, TextDocumentIdentifier document, Position position) {
-            var res = await s.Completion(new CompletionParams { textDocument = document, position = position }, CancellationToken.None);
-            DumpDetails(res);
-            if (res.items != null && res.items.Any()) {
-                var msg = string.Join(", ", res.items.Select(c => c.label).Ordered());
-                Assert.Fail("Completions were returned: " + msg);
-            }
         }
 
         public static async Task AssertSignature(Server s, TextDocumentIdentifier document, SourceLocation position, IReadOnlyCollection<string> contains, IReadOnlyCollection<string> excludes, string expr = null) {
